backend/services/image_infer.py
"""
Image Analysis Service
Analyzes facial expressions for emotional state using FER
Currently using placeholder logic - will be replaced with fer library or MobileNet
"""

import io
import random
from typing import Dict


# Facial emotion categories
FACIAL_EMOTIONS = ['angry', 'disgust', 'fear', 'happy', 'sad', 'surprise', 'neutral']

# Map facial emotions to stress levels
FACIAL_STRESS_MAP = {
    'angry': 0.75,
    'disgust': 0.65,
    'fear': 0.85,
    'happy': 0.15,
    'sad': 0.70,
    'surprise': 0.40,
    'neutral': 0.35
}


def analyze_image(image_bytes: bytes, filename: str) -> Dict:
    """
    Analyze facial expression for emotional state
    
    Args:
        image_bytes: Raw image file bytes
        filename: Original filename
    
    Returns:
        {
            'score': float (0-1, higher = more concerning),
            'bucket': str ('Low', 'Moderate', 'High'),
            'top_emotions': Dict[str, float] (top 3 emotions),
            'explain': {
                'all_emotions': Dict[str, float],
                'dominant_emotion': str,
                'face_detected': bool,
                'confidence': float
            }
        }
    """
    
    # Placeholder: Generate deterministic scores based on filename
    filename_hash = sum(ord(c) for c in filename)
    random.seed(filename_hash)
    
    # Simulate face detection (95% success rate)
    face_detected = random.random() > 0.05
    
    if not face_detected:
        return {
            "score": 0.5,  # Neutral when no face
            "bucket": "Moderate",
            "top_emotions": {"neutral": 1.0},
            "explain": {
                "all_emotions": {"neutral": 1.0},
                "dominant_emotion": "neutral",
                "face_detected": False,
                "confidence": 0.0,
                "message": "No face detected in image"
            }
        }
    
    # Simulate emotion detection
    emotions_prob = {}
    total = 0
    for emotion in FACIAL_EMOTIONS:
        prob = random.random()
        emotions_prob[emotion] = prob
        total += prob
    
    # Normalize probabilities
    for emotion in emotions_prob:
        emotions_prob[emotion] = round(emotions_prob[emotion] / total, 3)
    
    # Sort and get top 3 emotions
    sorted_emotions = sorted(emotions_prob.items(), key=lambda x: x[1], reverse=True)
    top_3 = dict(sorted_emotions[:3])
    dominant_emotion = sorted_emotions[0][0]
    confidence = sorted_emotions[0][1]
    
    # Calculate stress score
    stress_contribution = sum(
        emotions_prob[e] * FACIAL_STRESS_MAP[e]
        for e in emotions_prob
    )
    
    score = round(stress_contribution, 3)
    
    # Categorize
    if score < 0.33:
        bucket = "Low"
    elif score < 0.66:
        bucket = "Moderate"
    else:
        bucket = "High"
    
    return {
        "score": score,
        "bucket": bucket,
        "top_emotions": top_3,
        "explain": {
            "all_emotions": emotions_prob,
            "dominant_emotion": dominant_emotion,
            "face_detected": True,
            "confidence": confidence,
            "facial_landmarks": {
                "eyes": "detected",
                "mouth": "detected",
                "eyebrows": "detected"
            }
        }
    }


def aggregate_daily_images(image_results: list) -> Dict:
    """
    Aggregate multiple images from the same day into a daily visual mood index
    
    Args:
        image_results: List of image analysis results
    
    Returns:
        Aggregated score and emotion distribution
    """
    if not image_results:
        return {
            "score": 0.5,
            "dominant_emotion": "neutral",
            "emotion_trend": {}
        }
    
    # Average scores
    avg_score = sum(r["score"] for r in image_results) / len(image_results)
    
    # Aggregate emotions
    emotion_totals = {}
    for result in image_results:
        for emotion, prob in result["top_emotions"].items():
            emotion_totals[emotion] = emotion_totals.get(emotion, 0) + prob
    
    # Normalize
    total = sum(emotion_totals.values())
    emotion_trend = {
        e: round(v / total, 3) 
        for e, v in emotion_totals.items()
    }
    
    dominant_emotion = max(emotion_trend.items(), key=lambda x: x[1])[0]
    
    return {
        "score": round(avg_score, 3),
        "dominant_emotion": dominant_emotion,
        "emotion_trend": emotion_trend,
        "num_images": len(image_results)
    }


# ===== Future: Real Model Implementation =====
"""
def analyze_image_fer(image_bytes: bytes) -> Dict:
    # TODO: Replace with actual FER implementation
    
    from fer import FER
    import cv2
    import numpy as np
    from PIL import Image
    
    # Load image
    image = Image.open(io.BytesIO(image_bytes))
    image_array = np.array(image)
    
    # Convert RGB to BGR for OpenCV
    if len(image_array.shape) == 3 and image_array.shape[2] == 3:
        image_array = cv2.cvtColor(image_array, cv2.COLOR_RGB2BGR)
    
    # Detect emotions
    detector = FER(mtcnn=True)  # Use MTCNN for better face detection
    result = detector.detect_emotions(image_array)
    
    if not result:
        return no_face_response()
    
    # Get dominant emotion
    emotions = result[0]['emotions']
    
    # Calculate stress score
    score = sum(emotions[e] * FACIAL_STRESS_MAP[e] for e in emotions)
    
    return format_result(emotions, score)
"""
